refactor: extract thinking-agent registry from ollama_client
Split the Ollama-powered transcript reasoning code into a thin transport
layer (ollama_client.py) and a new thinking_agents.py that owns prompts,
model selection, response parsing, and a registry of Agent entries.

The two bespoke daemon-thread spawners in transcripts_server.py are
replaced with a generic orchestrator (_next_eligible_agent / _run_agent /
_run_agent_chain) that walks the AGENTS list and respects each agent's
depends_on and manifest_field. Adding a new agent (keyword extractor,
tagger, etc.) now requires appending one entry to AGENTS — no edits to
the orchestrator or transport needed.

CLAUDE.md gains a short "Local AI" section describing the seam so future
agents can find it quickly.

Behavior is preserved: summary + citations generation, regeneration
endpoints, in-flight tracking, and manifest schema are unchanged. All
542 tests pass; ruff + ty clean.
